app/etl/requirements.txt
python-dateutil==2.8.1
requests==2.27.1
lxml==4.6.5
xmltodict==0.12.0
Jinja2==2.11.1
pymongo==3.10.1
xxhash==1.4.3
motor==2.1.0
pydantic==1.6.2
pyproj==2.6.1.post1
SQLAlchemy==1.3.18
marshmallow-sqlalchemy==0.23.1
beautifulsoup4==4.9.1
pytz==2021.1
xmltodict==0.12.0
python-dotenv==0.21.0